Document how to compile and run tests on Windows
The first time compiling, we are faced with a linking error on Windows,
which is due to a dependency internally needing the access to the
WinPCAP Dev tools available.

After some search, this commit introduces documentation on how to avoid
the error.

I'm still not sure if the library would be necessary on runtime as a
.dll, but let's not get ahead of ourselves

Signed-off-by: Bruno Tavares <connect+github@bltavares.com>